Changed how connections are stored, in order for reconnecting to work properly and to handle disconnects. Program runs as normal but the new functionality, namely the reconnecting client needs testing, pushing this change to test with Emil.
